`launch v2`: move Plan types into their own package
