+++ internal/command/launch/plan/postgres.go
@@ -1,4 +1,4 @@
-package launch
+package plan
 
 import (
 	"context"
