+++ internal/command/launch/plan/redis.go
@@ -1,4 +1,4 @@
-package launch
+package plan
 
 import (
 	"context"
